Move UrlBuilder out to be a more general tool.
