+++ app/com/gu/identity/frontend/models/UrlBuilder.scala
@@ -1,0 +1,24 @@
+package com.gu.identity.frontend.models
+
+import play.api.mvc.Call
+
+object UrlBuilder {
+
+  private def encode = java.net.URLEncoder.encode(_: String, "UTF8")
+
+  def apply(baseUrl: String, params: Seq[(String, String)]): String =
+    params.headOption match {
+      case None => baseUrl
+      case _ => {
+        val paramString = params.map {
+          case (key, value) => s"$key=${encode(value)}"
+        }.mkString("&")
+
+        s"$baseUrl?$paramString"
+      }
+    }
+
+  def apply(call: Call, params: Seq[(String, String)]): String = {
+    apply(call.url, params)
+  }
+}
